pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-weekly-report"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.scripts]
swr     = "cli:main"
swr-mcp = "mcp_server:main"

[tool.setuptools]
py-modules = ["cli", "mcp_server", "config_manager"]
